Add options
I added the option to update all repos from remote (just a fetch, not a
pull).  I also added the option to get the status of the repos.  Only
the repos with a status message will show up in the list now.  The
default now only shows a list of all the repos in your $HOME directory.

The option logic is only set up for one argument.  Making it more robust
is on my TODO list.