Replace find statement with printf statement
I was getting an error when I ran this with the error settings in place.
It turns out that *find* was the source: it returns a 1 if you try to
access files and get "Permission denied" messages.

I first tried rearranging the error checking to be set during debug
only, but I didn't think that was the best solution.  I thought maybe a
*grep* statement would do the trick--which it did, but verrry slowly.  I
then tried an *echo* statement which was very fast and error-free (once I
figured out how to get globbing to work the way I wanted it to).

After reading a little about *printf*, I decided to go with that.